+++ src/components/features/auth/refresh-token/index.ts
@@ -1,1 +1,2 @@
-export { invalidateAccessTokenListener } from "./middleware";
+export {refreshTokenThunk} from "./thunk";
+
