+++ src/components/shared/api/actions.ts
@@ -5,3 +5,8 @@ export const invalidateAccessTokenAction = createAction(
   'RTK/Query/invalidateAccessToken'
 )
 
+export const refreshTokenAction = () =>{
+    const {refreshTokenThunk} = require("components/features/auth/refresh-token");
+    return refreshTokenThunk()
+}
+
