+++ src/components/shared/api/baseQueryWithReAuth.ts
@@ -10,7 +10,9 @@ import {
 
 import {baseQuery} from "./baseQuery";
 
-import {invalidateAccessTokenAction} from "./actions"
+import {refreshTokenAction} from "./actions"
+
+
 
 const AUTH_ERROR_CODES = new Set([403])
 
@@ -22,24 +24,13 @@ export async function baseQueryWithReAuth(
 ): Promise<QueryReturnValue<unknown, FetchBaseQueryError, FetchBaseQueryMeta>> {
     let result = await baseQuery(args, api, extraOptions)
 
-    /**
-     * 👇 ATTENTION: We can't use any thunk in direct mode,
-     * coz it's FSD Violation:
-     *
-     * api.dispatch(logoutThunk()) // 👎
-     *
-     * So we dispatch shared event `invalidateAccessToken`,
-     * which has subscribes via redux middleware in other layers.
-     * See example in @/features/authentication/InvalidateAccessToken
-     */
-
     if (
         typeof result?.error?.status === 'number'
         && AUTH_ERROR_CODES.has(result.error!.status)
     ){
-        console.log('sending invalidateAccessTokenAction')
+        // console.log('sending invalidateAccessTokenAction')
         // send refresh token to get new access token
-        await api.dispatch(invalidateAccessTokenAction())
+        await api.dispatch(refreshTokenAction())
 
         // retry with new access token
         result = await baseQuery(args, api, extraOptions)
